feat(frontend): @agenta/navigation — the sidebar's model leaves the app
