+++ web/packages/agenta-navigation/package.json
@@ -1,0 +1,37 @@
+{
+    "name": "@agenta/navigation",
+    "version": "0.1.0",
+    "private": true,
+    "sideEffects": false,
+    "main": "./src/index.ts",
+    "types": "./src/index.ts",
+    "scripts": {
+        "build": "tsc --noEmit",
+        "types:check": "tsc --noEmit",
+        "lint": "eslint --config ./eslint.config.mjs src/",
+        "check": "pnpm run types:check && pnpm run lint"
+    },
+    "exports": {
+        ".": "./src/index.ts"
+    },
+    "dependencies": {
+        "@agenta/entities": "workspace:../agenta-entities",
+        "@agenta/sessions": "workspace:../agenta-sessions",
+        "@agenta/shared": "workspace:../agenta-shared",
+        "@phosphor-icons/react": "^2.1.10",
+        "jotai-tanstack-query": "^0.11.0"
+    },
+    "peerDependencies": {
+        "@tanstack/react-query": ">=5.0.0",
+        "jotai": ">=2.0.0",
+        "react": ">=18.0.0"
+    },
+    "devDependencies": {
+        "@tanstack/react-query": "5.100.9",
+        "@types/node": "^20.19.20",
+        "@types/react": "^19.0.10",
+        "jotai": "^2.15.0",
+        "react": "^19.0.0",
+        "typescript": "^5.9.3"
+    }
+}
